requirements.txt
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.39
numpy==1.26.4
ollama==0.3.3
onnx==1.17.0
onnxruntime==1.19.2
openwakeword==0.6.0
protobuf==5.29.0rc1
PyAudio==0.2.14
pyserial==3.5
requests==2.32.3
sounddevice==0.5.1
soundfile==0.12.1
torch==2.5.0
torchaudio==2.5.0
tqdm==4.66.5
transformers==4.45.2